# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: api
  description: ''
  license:
    name: ''
  version: 0.1.0
servers:
- url: /api
paths:
...
  /admin/integration-events:
    get:
      tags:
      - admin
      operationId: list_admin_integration_events
      parameters:
      - name: provider
        in: query
        description: Optional integration provider filter
        required: false
        schema:
          type: string
        example: google
      - name: user_id
        in: query
        description: Optional Bike user id filter
        required: false
        schema:
          type: integer
          format: int32
        example: 12
      - name: activity_id
        in: query
        description: Optional activity id filter against event payload
        required: false
        schema:
          type: integer
          format: int32
        example: 42
      - name: import_id
        in: query
        description: Optional activity import id filter against event payload
        required: false
        schema:
          type: integer
          format: int32
        example: 17
      - name: limit
        in: query
        description: Maximum number of rows to return
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
        example: 20
      responses:
        '200':
          description: Recent integration events for administrators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IntegrationEventResponse'
              example:
              - connection_id: 5
                created_at: 2026-09-27T12:00:00Z
                event_type: sync_completed
                id: 31
                level: info
                message: Imported 3 activities
                payload:
                  imported_count: 3
                provider: strava
                user_id: 12
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearer_auth: []
...
  /integration-events/strava:
    get:
      tags:
      - strava
      operationId: list_strava_history
      responses:
        '200':
          description: Recent Strava integration history for the authenticated user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IntegrationEventResponse'
              example:
              - connection_id: 5
                created_at: 2026-09-27T12:00:00Z
                event_type: sync_completed
                id: 31
                level: info
                message: Imported 3 activities
                payload:
                  imported_count: 3
                provider: strava
                user_id: 12
        '401':
          description: Not authenticated
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
      - bearer_auth: []
...
components:
  schemas:
...
    ApiErrorResponse:
      type: object
      required:
      - message
      properties:
        errors:
          type:
          - object
          - 'null'
          additionalProperties:
            type: array
            items:
              type: string
          propertyNames:
            type: string
          example:
            title:
            - Title cannot be empty
        message:
          type: string
          example: Invalid request
        retry_at:
          type:
          - string
          - 'null'
          format: date-time
      example:
        errors:
          title:
          - Title cannot be empty
        message: Invalid request
...
    IntegrationEventResponse:
      type: object
      required:
      - id
      - provider
      - event_type
      - level
      - message
      - created_at
      properties:
        connection_id:
          type:
          - integer
          - 'null'
          format: int32
          example: 5
        created_at:
          type: string
          format: date-time
          example: 2026-09-27T12:00:00Z
        event_type:
          type: string
          example: sync_completed
        id:
          type: integer
          format: int32
          example: 31
        level:
          type: string
          example: info
        message:
          type: string
          example: Imported 3 activities
        payload: {}
        provider:
          type: string
          example: strava
        user_id:
          type:
          - integer
          - 'null'
          format: int32
          example: 12
      example:
        connection_id: 5
        created_at: 2026-09-27T12:00:00Z
        event_type: sync_completed
        id: 31
        level: info
        message: Imported 3 activities
        payload:
          imported_count: 3
        provider: strava
        user_id: 12
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...